+++ synapse/config/encryption_config.py
@@ -5,4 +5,4 @@ from cryptography.fernet import Fernet
 
 load_dotenv()
 
-cipher = Fernet(os.getenv("FERNET_KEY"))
+cipher = Fernet(os.getenv("FERNET_KEY", ""))